requirements.txt
fire==0.7.1
fastapi==0.117.1
uvicorn==0.37.0
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
Jinja2==3.1.6
wandb==0.22.0
